services/ml/requirements.txt
pep8==1.7.1
bcrypt==3.1.3
git+git://github.com/scikit-learn-contrib/boruta_py.git#egg=boruta
boruta
catboost==0.6.3
Flask==0.10.1
Flask-Cors==2.1.2
numpy==1.14.0
gunicorn==19.6.0
lifelines==0.13.0
lime==0.1.1.29
mongomock==3.6.0
pandas==0.22.0
PyJWT==1.4.2
pymongo==3.2.2
pystmark==0.4.5
requests==2.8.1
scikit-learn==0.19.1
scipy==1.0.0
dill==0.2.7.1
mandrill==1.0.57
